refactor: rename usecase mod name
