+++ src/usecase/es_list_task_usecase.rs
@@ -64,10 +64,10 @@ pub trait ListTaskUseCaseComponent {
 mod tests {
     use super::*;
     use crate::infra::sqlite::es_task_repository::TaskRepository;
-    use crate::usecase::add_es_task_usecase::{
+    use crate::usecase::es_add_task_usecase::{
         AddTaskUseCase, AddTaskUseCaseComponent, AddTaskUseCaseInput,
     };
-    use crate::usecase::close_es_task_usecase::{
+    use crate::usecase::es_close_task_usecase::{
         CloseTaskUseCase, CloseTaskUseCaseComponent, CloseTaskUseCaseInput,
     };
     use rusqlite::Connection;
